+++ libcove/lib/common.py
@@ -384,7 +384,11 @@ def get_schema_codelist_paths(
 
         if value.get("type") == "object":
             get_schema_codelist_paths(None, value, path, codelist_paths)
-        elif value.get("type") == "array" and value.get("items", {}).get("properties"):
+        elif (
+            value.get("type") == "array"
+            and isinstance(value.get("items"), dict)
+            and value.get("items").get("properties")
+        ):
             get_schema_codelist_paths(None, value["items"], path, codelist_paths)
 
     return codelist_paths
@@ -1157,7 +1161,11 @@ def _get_schema_deprecated_paths(
 
         if value.get("type") == "object":
             _get_schema_deprecated_paths(None, value, path, deprecated_paths)
-        elif value.get("type") == "array" and value.get("items", {}).get("properties"):
+        elif (
+            value.get("type") == "array"
+            and isinstance(value.get("items"), dict)
+            and value.get("items").get("properties")
+        ):
             _get_schema_deprecated_paths(None, value["items"], path, deprecated_paths)
 
     return deprecated_paths
@@ -1198,7 +1206,11 @@ def _get_schema_non_required_ids(
 
         if value.get("type") == "object":
             _get_schema_non_required_ids(None, value, path, id_paths)
-        elif value.get("type") == "array" and value.get("items", {}).get("properties"):
+        elif (
+            value.get("type") == "array"
+            and isinstance(value.get("items"), dict)
+            and value.get("items").get("properties")
+        ):
             has_list_merge = "wholeListMerge" in value and value.get("wholeListMerge")
             _get_schema_non_required_ids(
                 None,
@@ -1253,7 +1265,11 @@ def add_is_codelist(obj):
 
         if value.get("type") == "object":
             add_is_codelist(value)
-        elif value.get("type") == "array" and value.get("items", {}).get("properties"):
+        elif (
+            value.get("type") == "array"
+            and isinstance(value.get("items"), dict)
+            and value.get("items").get("properties")
+        ):
             add_is_codelist(value["items"])
 
     for value in obj.get("definitions", {}).values():