+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `get_schema_validation_errors` and therefore `common_checks_context` return more fields on each error dictionary, so that we can [replace the message with a translation in lib-cove-web](https://github.com/open-contracting/cove-ocds/issues/144)
 
+### Fixed
+
+- Don't error when the value for the `items` key in a JSON Schema is not a dict
+
 ## [0.21.0] - 2021-02-17
 
 ### Changed